Displays error when the resource is not found.
In the example with soft delete of users, it would be best if the user handles the case to redirect to the index page after deleting, but if it is not done, we display an error message instead of raising an error.

Remove some redundant imports because the BakcpexWeb html already imports it.